+++ src/Stratis.Bitcoin.IntegrationTests.Common/Extensions/CoreNodeExtensions.cs
@@ -1,4 +1,5 @@
-﻿using System.Linq;
+﻿using System.IO;
+using System.Linq;
 using NBitcoin;
 using Stratis.Bitcoin.Consensus;
 using Stratis.Bitcoin.Features.Consensus.Rules.CommonRules;
@@ -8,6 +9,14 @@ namespace Stratis.Bitcoin.IntegrationTests
 {
     public static class CoreNodeExtensions
     {
+        public static void AppendToConfig(this CoreNode node, string configKeyValueItem)
+        {
+            using (StreamWriter sw = File.AppendText(node.Config))
+            {
+                sw.WriteLine(configKeyValueItem);
+            }
+        }
+
         public static Money GetProofOfWorkRewardForMinedBlocks(this CoreNode node, int numberOfBlocks)
         {
             var coinviewRule = node.FullNode.NodeService<IConsensusRuleEngine>().GetRule<CoinViewRule>();
