+++ src/Stratis.Bitcoin.IntegrationTests.Common/Runners/StratisBitcoinPosRunner.cs
@@ -17,15 +17,21 @@ namespace Stratis.Bitcoin.IntegrationTests.Common.Runners
 {
     public sealed class StratisBitcoinPosRunner : NodeRunner
     {
-        public StratisBitcoinPosRunner(string dataDir, Network network, string agent = "StratisBitcoin")
+        private readonly bool isGateway;
+
+        public StratisBitcoinPosRunner(string dataDir, Network network, string agent = "StratisBitcoin", bool isGateway = false)
             : base(dataDir, agent)
         {
             this.Network = network;
+            this.isGateway = isGateway;
         }
 
         public override void BuildNode()
         {
             var settings = new NodeSettings(this.Network, ProtocolVersion.PROVEN_HEADER_VERSION, this.Agent, args: new string[] { "-conf=stratis.conf", "-datadir=" + this.DataFolder });
+
+            if (this.isGateway)
+                settings.MinProtocolVersion = ProtocolVersion.ALT_PROTOCOL_VERSION;
 
             var builder = new FullNodeBuilder()
                 .UseNodeSettings(settings)
