+++ src/Stratis.Features.FederatedPeg.IntegrationTests/Utils/SidechainNodeBuilder.cs
@@ -5,7 +5,6 @@ using Stratis.Bitcoin.Configuration;
 using Stratis.Bitcoin.Features.PoA;
 using Stratis.Bitcoin.Features.PoA.IntegrationTests.Common;
 using Stratis.Bitcoin.IntegrationTests.Common.EnvironmentMockUpHelpers;
-using Stratis.Bitcoin.IntegrationTests.Common.Runners;
 
 namespace Stratis.Features.FederatedPeg.IntegrationTests.Utils
 {
@@ -61,27 +60,14 @@ namespace Stratis.Features.FederatedPeg.IntegrationTests.Utils
             return node;
         }
 
-        public CoreNode CreateMainChainFederationNode(Network network, Network counterChainNetwork, NodeConfigParameters parameters = null)
+        public CoreNode CreateMainChainFederationNode(Network network, Network counterChainNetwork)
         {
             string agentName = $"mainfed{Interlocked.Increment(ref agentCount)}";
             string dataFolder = this.GetNextDataFolderName(agentName);
 
-            CoreNode node = this.CreateNode(new MainChainFederationNodeRunner(dataFolder, agentName, network, counterChainNetwork), "stratis.conf", configParameters: parameters);
+            CoreNode node = this.CreateNode(new MainChainFederationNodeRunner(dataFolder, agentName, network, counterChainNetwork), "stratis.conf");
 
             return node;
-        }
-
-        public CoreNode CreateMainnetStratisXNode(string version = "2.0.0.5", bool useCookieAuth = false, NodeConfigParameters parameters = null)
-        {
-            if (parameters == null)
-            {
-                parameters = new NodeConfigParameters();
-            }
-            parameters.Add("regtest", "0");
-            parameters.Add("server", "0");
-
-            string stratisDPath = GetStratisXPath(version);
-            return this.CreateNode(new StratisXRunner(this.GetNextDataFolderName(), stratisDPath), "stratis.conf", useCookieAuth, parameters);
         }
     }
 }
